SongEditor: split into different files
instead of having `SongEditor`, `SongEditorPatternEditor`, and `SongEditorPositionRuler` reside in the same file, all of them are moved to a dedicated file.

In addition, caching of the `Hydrogen` and `AudioEngine` pointers was removed and all three classes were hardened against no song being present
